+++ frontend/app/components/modules/PaymentPay/PaymentPay.tsx
@@ -2,56 +2,88 @@ import Image from 'next/image'
 import { BigInput } from '@/elements/BigInput/BigInput'
 import { Button } from '@/elements/Button/Button'
 import { PinInput } from '@/elements/PinInput/PinInput'
-import { Color } from '@/types/elements'
+import { Color, PaymentAccountInfo } from '@/types/elements'
 import { PIN_DIGIT_LENGTH } from 'app/constants/amounts'
-import { useState } from 'react'
+import { Dispatch, SetStateAction, useEffect, useState } from 'react'
 import { Cam } from '../Cam/Cam'
 import styles from './PaymentPay.module.scss'
-import { checkLimit, pay } from '@/api/paymentAPI'
+import { checkAccount, pay, resetBalance } from '@/api/paymentAPI'
 import { getImageFromLocalStorage } from '@/utils/localStorage/localStorage'
 import { Spinner } from '@/elements/Spinner/Spinner'
 import { FACE_MATCH_LIVENESS_SNAPSHOT } from 'app/constants/localStorage'
 import { CustomError } from 'app/errors/CustomError'
+import { formatMoney } from '@/utils/utils'
 
-type Props = { sessionId: string; amount: number; afterPay: () => void }
+type Props = {
+  sessionId: string
+  amount: number
+  paymentAccountInfo: PaymentAccountInfo
+  setPaymentAccountInfo: Dispatch<
+    SetStateAction<PaymentAccountInfo | undefined>
+  >
+  afterPay: () => void
+  backToCatalog: () => void
+  backToStart: () => void
+}
 
-export const PaymentPay = ({ sessionId, amount, afterPay }: Props) => {
+export const PaymentPay = ({
+  sessionId,
+  amount,
+  paymentAccountInfo,
+  setPaymentAccountInfo,
+  afterPay,
+  backToCatalog,
+  backToStart
+}: Props) => {
   const [step, setStep] = useState(1)
   const [pinCode, setPinCode] = useState('')
   const [phone, setPhone] = useState('')
   const [isPinRequired, setIsPinRequired] = useState(true)
-  const [user, setUser] = useState('User')
-  const [balance, setBalance] = useState(0)
   const [isSuccess, setIsSuccess] = useState(false)
+  const [isBalanceSufficient, setIsBalanceSufficient] = useState(false)
   const [message, setMessage] = useState('')
   const [phoneError, setPhoneError] = useState('')
   const [isLoading, setIsLoading] = useState(false)
   const [httpCode, setHttpCode] = useState(500)
 
-  const resolveCheckLimit = async (
-    session_id: string,
-    phone: string,
-    amount: number
-  ) => {
+  useEffect(() => {
+    setIsBalanceSufficient(paymentAccountInfo.balance - amount >= 0)
+    setIsPinRequired(
+      paymentAccountInfo.minimum_payment < amount ||
+        paymentAccountInfo.have_twin
+    )
+  }, [
+    amount,
+    paymentAccountInfo.balance,
+    paymentAccountInfo.have_twin,
+    paymentAccountInfo.minimum_payment
+  ])
+
+  const checkPhone = () => {
+    if (paymentAccountInfo.phone !== phone) {
+      setPhoneError('Your phone number does not match with the registered one')
+    } else {
+      setPhoneError('')
+      setStep(2)
+    }
+  }
+
+  const resolveResetBalance = async (sessionId: string) => {
     try {
-      const res = await checkLimit(session_id, phone, amount)
-      if (res?.ok) {
-        setIsPinRequired(res.data[0].is_limit || res.data[0].have_twin)
-        setUser(res.data[0].full_name)
-        setBalance(res.data[0].balance)
-        setPhoneError('')
-      }
+      await resetBalance(sessionId)
+      const res = await checkAccount(sessionId)
+
+      setPaymentAccountInfo(res?.data)
+      backToCatalog()
     } catch (e) {
       if (e instanceof CustomError) {
+        setHttpCode(e.statusCode)
         switch (e.statusCode) {
-          case 400:
-            setPhoneError(e.message)
-            setStep(1)
-            break
-          case 401:
-            // TODO ADD OPEN MODAL
+          case 404:
+            backToStart()
             break
           default:
+            console.error(e)
             break
         }
       } else {
@@ -91,6 +123,7 @@ export const PaymentPay = ({ sessionId, amount, afterPay }: Props) => {
             break
           default:
             console.error(e)
+            break
         }
       } else {
         setIsSuccess(false)
@@ -115,8 +148,7 @@ export const PaymentPay = ({ sessionId, amount, afterPay }: Props) => {
             <Button
               color={Color.Primary}
               onClick={() => {
-                setStep(2)
-                resolveCheckLimit(sessionId, phone, amount)
+                checkPhone()
               }}>
               Next Step
             </Button>
@@ -152,35 +184,71 @@ export const PaymentPay = ({ sessionId, amount, afterPay }: Props) => {
       )}
       {step === 4 && (
         <div className={styles.resultWrapper}>
-          <h2>{`Hello, ${user}`}</h2>
+          <h2>{`Hello, ${paymentAccountInfo.full_name}`}</h2>
 
           <p className={styles.balance}>
-            You have to pay <strong>{amount}</strong>
+            You have to pay <strong>{`IDR ${formatMoney(amount)}`}</strong>
           </p>
           <p className={styles.balance}>
-            Current Balance <strong>{balance}</strong>
+            Current Balance{' '}
+            <strong>{`IDR ${formatMoney(paymentAccountInfo.balance)}`}</strong>
           </p>
-          <p className={styles.balance}>
-            Remaining Balance <strong>{balance - amount}</strong>
-          </p>
+          {isBalanceSufficient ? (
+            <>
+              <p className={styles.balance}>
+                Remaining Balance{' '}
+                <strong>
+                  {`IDR ${formatMoney(paymentAccountInfo.balance - amount)}`}
+                </strong>
+              </p>
 
-          <Button
-            type="button"
-            color={Color.Primary}
-            onClick={() => {
-              setStep(5)
-              resolvePay(
-                sessionId,
-                phone,
-                pinCode,
-                amount,
-                getImageFromLocalStorage(FACE_MATCH_LIVENESS_SNAPSHOT, () =>
-                  setStep(2)
-                )
-              )
-            }}>
-            Confirm
-          </Button>
+              <Button
+                type="button"
+                color={Color.Primary}
+                onClick={() => {
+                  setStep(5)
+                  resolvePay(
+                    sessionId,
+                    phone,
+                    pinCode,
+                    amount,
+                    getImageFromLocalStorage(FACE_MATCH_LIVENESS_SNAPSHOT, () =>
+                      setStep(2)
+                    )
+                  )
+                }}>
+                Confirm
+              </Button>
+            </>
+          ) : (
+            <>
+              <p className={styles.balance}>
+                <strong>Sorry, your balance is not enough</strong>
+              </p>
+              <div className={styles.btnGroup}>
+                <Button
+                  type="button"
+                  color={Color.Primary}
+                  onClick={() => backToCatalog()}>
+                  Back to Catalog
+                </Button>
+                <Button
+                  type="button"
+                  color={Color.Primary}
+                  onClick={() => resolveResetBalance(sessionId)}>
+                  Reset Balance
+                </Button>
+                <Button
+                  type="button"
+                  color={Color.Primary}
+                  onClick={() => {
+                    afterPay()
+                  }}>
+                  Finish
+                </Button>
+              </div>
+            </>
+          )}
         </div>
       )}
       {step === 5 &&