+++ CHANGELOG.md
@@ -7,13 +7,18 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - [BACKEND] Add HTTP Requester for OCR Receipt Recognition Innovation
-- [BACKEND] Add HTTP Requester for OCR Receipt Recognition Innovation
 - [BACKEND] Add API Reset Balance Account on Face Payment
+- [FRONTEND] Add current balance to catalog browse step
+- [FRONTEND] Add feature to remove local storage if face payment face analytics failed
+- [FRONTEND] Add back to catalog feature if balance insufficient
+- [FRONTEND] Add reset balance feature
 - [FRONTEND] Add shopping balance in catalog
 - [FRONTEND] Make Big Input Component responsive
 
 ### Changed
 
+- [FRONTEND] Change pin required checking to client side
+- [FRONTEND] Change phone checking to client side
 - [FRONTEND] Error handling mechanism
 - [FRONTEND] Revamp landing page
 - [FRONTEND] Fix Camera not full on mobile
@@ -28,6 +33,7 @@ All notable changes to this project will be documented in this file.
 
 ### Removed
 
+- [FRONTEND] Remove check-phone api handler
 - [BACKEND] Delete API Check Limit on Face Payment
 
 ## [0.3.0 - 17/11/2021]